+++ src/packages/map/internalTestSupport.ts
@@ -1,0 +1,11 @@
+// SPDX-FileCopyrightText: 2023 Open Pioneer project (https://github.com/open-pioneer)
+// SPDX-License-Identifier: Apache-2.0
+
+/**
+ * Exported for the `@open-pioneer/map-test-utils` package.
+ *
+ * This API is unstable.
+ *
+ * @internal
+ */
+export { MapRegistryImpl } from "./MapRegistryImpl";