+++ src/packages/map-navigation/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @open-pioneer/map-navigation
 
+## 0.4.2
+
+### Patch Changes
+
+-   4140646: Update trails dependencies
+-   4140646: Update to react 18.3.1
+-   f2912be: Added more context to some button labels and tooltips for better usability.
+-   81bc7da: Update trails dependencies
+-   2c092dc: Update dependencies
+-   Updated dependencies [4140646]
+-   Updated dependencies [4140646]
+-   Updated dependencies [b5bb7a1]
+-   Updated dependencies [81bc7da]
+-   Updated dependencies [2c092dc]
+-   Updated dependencies [4140646]
+    -   @open-pioneer/react-utils@0.2.3
+    -   @open-pioneer/map@0.5.1
+
 ## 0.4.1
 
 ### Patch Changes